requirements.txt
playwright==1.47.0
python-dotenv==1.0.1
PyMySQL==1.1.1
simple-justwatch-python-api 
httpx